move internal feed model to separate file
